pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spectral-scripts"
version = "0.1.0"
description = "Spectral analysis of OCR confusion matrices for script comparison"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Bao" }]

dependencies = [
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "rich>=13.0.0",
    "statsmodels>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]
ocr-tesseract = [
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
]
ocr-trocr = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
]
ocr-all = [
    "pytesseract>=0.3.10",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "easyocr>=1.7.0",
]

[project.scripts]
spectral-analyze = "spectral_scripts.cli:main"
run-ocr-pipeline = "spectral_scripts.ocr_pipeline.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/spectral_scripts"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=spectral_scripts --cov-report=term-missing"